[feature] Add commit options parsing
