+++ GitGudCLI/Options/CommitOptions.cs
@@ -5,7 +5,10 @@ namespace GitGudCLI.Options
 	[Verb("commit", true, HelpText = "Simplify commiting")]
 	public class CommitOptions
 	{
-		[Value(0, HelpText = "The commit message")] 
+		[Value(0, HelpText = "The commit mode (quickadd/qa, commit/c, fullcommit/fc)")] 
+		public string Mode { get; set; }
+		
+		[Value(1, HelpText = "The commit message")] 
 		public string Message { get; set; }
 	}
 }
