+++ GitGudCLI/Program.cs
@@ -1,5 +1,7 @@
 ﻿using System;
+using CommandLine;
 using ConsoleHelper;
+using GitGudCLI.Options;
 using GitGudCLI.Utils;
 
 namespace GitGudCLI
@@ -10,41 +12,20 @@ namespace GitGudCLI
 
 		private static void Main(string[] args)
 		{
+			_gitHelper = new GitHelper();
+			
+			Parser.Default.ParseArguments<CommitOptions>(args)
+				.WithParsed<CommitOptions>(options => CommitCommands.Run(options, _gitHelper));
+				//.WithNotParsed(errors => ...);
+			
 			if (args.Length == 0)
 			{
 				Console.WriteLine("HELP_TEXT");
 				return;
 			}
 
-			_gitHelper = new GitHelper();
-
 			switch (args[0])
 			{
-				case "commitadd":
-				case "ca":
-					CLIMethods.QuickCommit(args.Length >= 2 ? args[1] : string.Empty, true, _gitHelper);
-					break;
-
-				case "commit":
-				case "c":
-					CLIMethods.QuickCommit(args.Length >= 2 ? args[1] : string.Empty, false, _gitHelper);
-					break;
-
-				case "fullcommit":
-				case "fc":
-					CLIMethods.FullCommit(args.Length >= 2 ? args[1] : string.Empty, _gitHelper);
-					break;
-
-				case "commitlint":
-				case "cmli":
-					CLIMethods.ValidateCommit(args.Length >= 2 ? args[1] : string.Empty);
-					break;
-
-				case "commitgen":
-				case "cmgn":
-					CLIMethods.GenerateValidCommit(args.Length >= 2 ? args[1] : string.Empty);
-					break;
-
 				case "changelog":
 				case "chlg":
